crypto-regulator-checker/backend/requirements.txt
# Core dependencies
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
pydantic-settings==2.1.0
python-dotenv>=0.19.0,<0.20.0
httpx>=0.18.2,<0.19.0

# Database and caching
redis==5.0.1
cachetools==5.3.2

# ML and embeddings
numpy==1.26.3
sentence-transformers==2.5.1
chromadb==0.4.22
google-generative-ai==0.3.2

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest>=6.2.5,<6.3.0
pytest-asyncio>=0.15.1,<0.16.0
pytest-cov==4.1.0
pytest-mock==3.12.0
aioresponses==0.7.6

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.1

# Security
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.1.0

# Database
sqlalchemy>=1.4.0,<1.5.0
alembic>=1.7.0,<1.8.0
psycopg2-binary>=2.9.0,<3.0.0
sqlalchemy-utils==0.41.2

# Additional dependencies
aiofiles>=0.7.0,<0.8.0
python-magic>=0.4.27,<0.5.0
requests>=2.26.0,<2.27.0 